apple/t2: refactor kernel package
This commit refactors the kernel package in this profile, allowing
end-users to use the latest kernel package from nixpkgs without much
maintenance.

Because the t2linux project has a lot of patches which changes
periodically, including them in the repository directly is a maintenance
burden. Instead, this patch comes with an update script and a JSON file,
containing all of the patches to be applied.

Signed-off-by: Cassie Cheung <me@soopy.moe>